Add files via upload
In this code, the user is prompted to enter the total number of candies produced, which is stored in the variable candies. The program then calculates the number of boxes by dividing the total number of candies by the product of 15 and 24 using integer division (//). The number of bags is calculated by first dividing the total number of candies by 15 using integer division, and then taking the remainder when this result is divided by 24 using the modulo operator (%). The remaining number of candies is calculated by taking the remainder when the total number of candies is divided by 15 using the modulo operator. Finally,